standard-setup
Standardize the project setup and different variables
