+++ .streamlit/secrets.example.toml
@@ -1,17 +1,22 @@
-# Configuration Variables
+ADMIN_KEY=""
+
+[general]
 MONGODB_URI=""
 UMAMI_WEBSITE_ID=""
-ADMIN_KEY=""
-SUPER_ADMIN_KEY=""
-GEMINI_API_KEY=""
+YOUTUBE_VIDEO_ID=""
+
+[api_key]
 NEWS_API_KEY=""
 TMDB_API_KEY=""
 NASA_API_KEY=""
+GEMINI_API_KEY=""
+OpenAI_API_KEY=""
 WEATHER_API_KEY=""
 YOUTUBE_API_KEY=""
-OpenAI_API_KEY=""
 UBERDUCK_API_KEY=""
+SPOONACULAR_API_KEY=""
 
-[google_drive]
+[cloud]
+[movie_recommendation_model]
 SIMILARITY=""
 MOVIES_LIST=""
