+++ src/apps/public/home.py
@@ -2,9 +2,10 @@ import streamlit as st
 
 def home():
 	st.title("Welcome to Jarvis - Your Virtual AI Assistant!")
-	st.image('assets/intro.gif', caption='Empower Your Digital Life with Jarvis', use_column_width=True)
+	st.toast("Welcome to Jarvis!", icon="👋")
+	st.video('assets/video.mp4', loop=True, autoplay=True, muted=True)
 
-    st.markdown("""
+	st.markdown("""
 	## What is Jarvis?
 	**Jarvis** is a powerful Python-based AI assistant designed to streamline your daily tasks through simple voice commands. Whether you're navigating the web, managing your media, or automating routine processes, Jarvis is here to help.
 	""")
@@ -20,7 +21,7 @@ def home():
 	""")
 
 	with st.expander("## See Jarvis in Action"):
-		st.video(f'https://www.youtube.com/watch?v={st.secrets["YOUTUBE_VIDEO_ID"]}', start_time=0)
+		st.video(f'https://www.youtube.com/watch?v={st.secrets["general"]["YOUTUBE_VIDEO_ID"]}', start_time=0)
 
 	st.markdown("""
 	## Learn More and Get Started
@@ -35,5 +36,3 @@ def home():
 	---
 	**Jarvis** is continually evolving with new features and improvements. Stay tuned for updates and feel free to contribute to its development.
 	""")
-
-home()
